Export other errors to error.rs
