+++ src/lib.rs
@@ -1,7 +1,6 @@
-use std::sync::mpsc::{SendError, Sender};
-use thiserror::Error;
+use std::sync::mpsc::Sender;
 use tungstenite::connect;
-use url::{ParseError, Url};
+use url::Url;
 
 use crate::handlers::error::*;
 use crate::types::{MessageFields, Session, TwitchMessage};
@@ -11,79 +10,6 @@ pub use serde_json::from_str as parse_message;
 
 pub mod handlers;
 pub mod types;
-
-#[derive(Error, Debug)]
-pub enum SessionErr {
-    #[error("error parsing url: {0}")]
-    Parse(ParseError),
-    #[error("connection error: {0}")]
-    Connect(tungstenite::Error),
-}
-
-impl From<ParseError> for EventSubErr {
-    fn from(err: ParseError) -> Self {
-        EventSubErr::Session(SessionErr::Parse(err))
-    }
-}
-
-impl From<tungstenite::Error> for SessionErr {
-    fn from(err: tungstenite::Error) -> Self {
-        SessionErr::Connect(err)
-    }
-}
-
-pub fn get_session(url: Option<&str>) -> Result<Session, EventSubErr> {
-    let to_parse;
-    if let Some(url) = url {
-        to_parse = url;
-    } else {
-        to_parse = "wss://eventsub-beta.wss.twitch.tv/ws";
-    }
-    let (socket, _) = connect(Url::parse(to_parse)?)?;
-    Ok(Session::new(socket))
-}
-
-#[derive(Error, Debug)]
-pub enum EventSubErr {
-    #[error("general handler error: {0}")]
-    GeneralHandler(HandlerErr),
-    #[error("socket error: {0}")]
-    Socket(tungstenite::Error),
-    #[error("session error: {0}")]
-    Session(SessionErr),
-    #[error("error sending through channel: {0}")]
-    Sending(SendError<TwitchMessage>),
-}
-
-impl From<EventSubErr> for String {
-    fn from(err: EventSubErr) -> String {
-        err.to_string()
-    }
-}
-
-impl From<HandlerErr> for EventSubErr {
-    fn from(err: HandlerErr) -> Self {
-        EventSubErr::GeneralHandler(err)
-    }
-}
-
-impl From<SessionErr> for EventSubErr {
-    fn from(err: SessionErr) -> Self {
-        EventSubErr::Session(err)
-    }
-}
-
-impl From<tungstenite::Error> for EventSubErr {
-    fn from(err: tungstenite::Error) -> Self {
-        EventSubErr::Socket(err)
-    }
-}
-
-impl From<SendError<TwitchMessage>> for EventSubErr {
-    fn from(err: SendError<TwitchMessage>) -> Self {
-        EventSubErr::Sending(err)
-    }
-}
 
 pub fn event_handler(
     session: &mut Session,
@@ -110,6 +36,17 @@ pub fn event_handler(
 
         session.handled.push(message_id.to_owned());
     }
+}
+
+pub fn get_session(url: Option<&str>) -> Result<Session, EventSubErr> {
+    let to_parse;
+    if let Some(url) = url {
+        to_parse = url;
+    } else {
+        to_parse = "wss://eventsub-beta.wss.twitch.tv/ws";
+    }
+    let (socket, _) = connect(Url::parse(to_parse)?)?;
+    Ok(Session::new(socket))
 }
 
 #[cfg(test)]
